refactor: rip out malloc_v1, fix makefile issue and add wip allocate_huge_chunk method

+++ src/my_malloc.c
@@ -1,4 +1,4 @@
-#include "../../inc/main.h" 
+#include "../inc/main.h" 
 
 void* my_malloc(size_t size) {
     if (g_pagemap == NULL) {
@@ -20,11 +20,11 @@ void* my_malloc(size_t size) {
 
 void* search_heap(size_t size, int page_type) {
     switch (page_type) {
-        case 1: return search_fast_cache(size);
-        case 2: return search_cache(size, page_type);
-        case 3: return search_cache(size, page_type);
-        case 4: return allocate_huge_chunk(size);
-        default: break;
+    case 1: return search_fast_cache(size);
+    case 2: return search_cache(size, page_type);
+    case 3: return search_cache(size, page_type);
+    case 4: return allocate_huge_chunk(size);
+    default: break;
     }
 
     printf("returning NULL\n");
@@ -34,8 +34,8 @@ void* search_heap(size_t size, int page_type) {
 int get_page_type(size_t size) {
     if (size < FAST_MAX_CHUNK_SIZE) return 1;
     else if (size < SMALL_MAX_CHUNK_SIZE) return 2;
-    else if (size < LARGE_MAX_CHUNK_SIZE) return 3;
-    else if (size < BASE_HEAP_SIZE) return 4;
+    else if (size < (size_t)LARGE_MAX_CHUNK_SIZE) return 3;
+    else if (size < (size_t)BASE_HEAP_SIZE) return 4;
     else return 0;
 }
